refactor: split runners media_filter + audio_dsp into focused modules
