+++ runners/audio_dsp.py
@@ -6,7 +6,6 @@ from .media import (
     _decode_audio_to_array, _AUDIO_RATE,
 )
 from .media_filter import make_progress
-from .roseus_lut import ROSEUS_LUT
 
 
 def _write_wav(arr, out_codec: str = 'wav', metadata: dict | None = None) -> str:
@@ -271,224 +270,8 @@ def segment_export(view_url: str, segments: list | None = None,
     return {'files': files, 'count': len(files)}
 
 
-_GRAY_LUT = tuple((v, v, v) for v in range(256))
-
-
-def _lut(colormap: str):
-    return ROSEUS_LUT if colormap == 'roseus' else _GRAY_LUT
-
-
-def render_waveform_image(view_url: str, width: int = 1200, height: int = 480,
-                          show_rms: bool = True, show_clipping: bool = True,
-                          db_axis: bool = False) -> str:
-    import numpy as np
-    from PIL import Image
-
-    arr = _decode_audio_to_array(localize(view_url))
-    if arr.shape[1] == 0:
-        raise RuntimeError("waveform: source has no audio")
-    width = min(4096, max(240, int(width)))
-    height = min(2048, max(120, int(height)))
-
-    mono = arr.mean(axis=0)
-    n = mono.shape[0]
-    block = max(1, n // width)
-    ncols = min(width, (n + block - 1) // block)
-    padded = np.zeros(ncols * block, dtype=np.float32)
-    padded[:n] = mono[:ncols * block]
-    cols = padded.reshape(ncols, block)
-    vmax = cols.max(axis=1)
-    vmin = cols.min(axis=1)
-    rms = np.sqrt((cols ** 2).mean(axis=1))
-    clip_cols = (np.abs(arr[:, :ncols * block])
-                 .reshape(2, ncols, block).max(axis=(0, 2))) >= (32766.0 / 32768.0)
-
-    def to_y(v):
-        if db_axis:
-            db = 20.0 * np.log10(np.maximum(np.abs(v), 1e-6))
-            mag = np.clip((db + 60.0) / 60.0, 0.0, 1.0) * np.sign(v)
-        else:
-            mag = np.clip(v, -1.0, 1.0)
-        return ((1.0 - mag) * 0.5 * (height - 1)).astype(int)
-
-    img = np.zeros((height, width, 3), dtype=np.uint8)
-    img[:, :] = (20, 20, 32)
-    mid = height // 2
-    img[mid, :ncols] = (60, 60, 90)
-
-    y_hi, y_lo = to_y(vmax), to_y(vmin)
-    yr_hi, yr_lo = to_y(rms), to_y(-rms)
-    for x in range(ncols):
-        img[min(y_hi[x], y_lo[x]):max(y_hi[x], y_lo[x]) + 1, x] = (86, 86, 149)
-        if show_rms:
-            img[min(yr_hi[x], yr_lo[x]):max(yr_hi[x], yr_lo[x]) + 1, x] = (130, 130, 200)
-        if show_clipping and clip_cols[x]:
-            img[:, x] = (255, 40, 40)
-
-    out = fresh_output_path('.png', subfolder='comfytv/audio')
-    Image.fromarray(img, 'RGB').save(str(out), 'PNG')
-    return path_to_view_url(out)
-
-
-def render_spectrogram_image(view_url: str, width: int = 1200,
-                             height: int = 480, scale: str = 'log',
-                             colormap: str = 'roseus', range_db: float = 80.0,
-                             gain_db: float = 20.0,
-                             freq_gain_dbpoct: float = 0.0) -> str:
-    import numpy as np
-    from PIL import Image
-
-    if scale not in ('linear', 'log', 'mel'):
-        raise RuntimeError(f"spectrogram: unknown scale {scale!r}")
-    arr = _decode_audio_to_array(localize(view_url))
-    if arr.shape[1] == 0:
-        raise RuntimeError("spectrogram: source has no audio")
-    width = min(4096, max(240, int(width)))
-    height = min(2048, max(120, int(height)))
-
-    mono = arr.mean(axis=0)
-    nfft = 2048
-    hop = max(1, (mono.shape[0] - nfft) // width)
-    nframes = min(width, max(1, 1 + (mono.shape[0] - nfft) // hop))
-    win = np.hanning(nfft).astype(np.float32)
-    idx = (np.arange(nframes)[:, None] * hop + np.arange(nfft)[None, :])
-    frames = mono[np.clip(idx, 0, mono.shape[0] - 1)] * win
-    power = np.abs(np.fft.rfft(frames, axis=1)) ** 2
-    wss = 4.0 / (win.sum() ** 2)
-    db = 10.0 * np.log10(np.maximum(power * wss, 1e-12))
-
-    freqs = np.fft.rfftfreq(nfft, 1.0 / _AUDIO_RATE)
-    if freq_gain_dbpoct:
-        db = db + freq_gain_dbpoct * np.log2(np.maximum(freqs, 1.0) / 1000.0)[None, :]
-
-    fmin, fmax = 20.0, _AUDIO_RATE / 2.0
-    rows = np.arange(height)
-    frac = 1.0 - rows / (height - 1)
-    if scale == 'linear':
-        target = frac * fmax
-    elif scale == 'log':
-        target = fmin * (fmax / fmin) ** frac
-    else:
-        def to_mel(f):
-            return 2595.0 * np.log10(1.0 + f / 700.0)
-
-        def from_mel(m):
-            return 700.0 * (10.0 ** (m / 2595.0) - 1.0)
-        target = from_mel(frac * to_mel(fmax))
-    bins = np.clip(np.searchsorted(freqs, target), 0, freqs.shape[0] - 1)
-
-    grid = db[:, bins].T
-    bright = np.clip((grid + float(gain_db) + float(range_db)) / float(range_db),
-                     0.0, 1.0)
-    lut = np.array(_lut(colormap), dtype=np.uint8)
-    img = lut[(bright * 255).astype(np.uint8)]
-    if nframes < width:
-        pad = np.zeros((height, width - nframes, 3), dtype=np.uint8)
-        pad[:, :] = lut[0]
-        img = np.concatenate([img, pad], axis=1)
-
-    out = fresh_output_path('.png', subfolder='comfytv/audio')
-    Image.fromarray(img, 'RGB').save(str(out), 'PNG')
-    return path_to_view_url(out)
-
-
-def convolve_ir(view_url: str, ir_url: str, wet: float = 1.0,
-                dry: float = 0.0, normalize: bool = True,
-                out_codec: str = 'wav', progress=None) -> str:
-    import numpy as np
-    from scipy.signal import fftconvolve
-
-    x = _decode_audio_to_array(localize(view_url))
-    ir = _decode_audio_to_array(localize(ir_url))
-    if x.shape[1] == 0:
-        raise RuntimeError("convolve: source has no audio")
-    if ir.shape[1] == 0:
-        raise RuntimeError("convolve: IR has no audio")
-
-    total = x.shape[1] + ir.shape[1] - 1
-    report = make_progress(progress, 2, "convolve")
-    out = np.zeros((2, total), dtype=np.float32)
-    for c in range(2):
-        out[c] = fftconvolve(x[c], ir[c], mode='full').astype(np.float32)
-        report(c + 1, text=f"convolve {c + 1}/2")
-    out *= float(wet)
-    if float(dry):
-        out[:, :x.shape[1]] += float(dry) * x
-    if normalize:
-        peak = float(np.abs(out).max())
-        if peak > 0.99:
-            out *= 0.99 / peak
-    return _write_wav(np.clip(out, -1.0, 1.0), out_codec)
-
-
-def _ess_signals(duration_s: float, fmin: float, fmax: float, amp: float,
-                 fade_in_s: float = 0.1, fade_out_s: float = 0.03):
-    import numpy as np
-
-    rate = _AUDIO_RATE
-    n_pre = int(rate * fade_in_s)
-    n_sin = int(rate * duration_s)
-    n_end = int(rate * fade_out_s)
-    n = n_pre + n_sin + n_end
-
-    a = np.log(fmax / fmin) / n_sin
-    b = fmin / (a * rate)
-    r = 4.0 * a * a / amp
-
-    i = np.arange(n, dtype=np.float64)
-    j = n - i - 1
-    gain = np.ones(n)
-    if n_pre > 0:
-        gain[:n_pre] = np.sin(0.5 * np.pi * i[:n_pre] / n_pre)
-    if n_end > 0:
-        tail = j < n_end
-        gain[tail] = np.sin(0.5 * np.pi * j[tail] / n_end)
-
-    d = b * np.exp(a * (i - n_pre))
-    p = d - b
-    x = gain * np.sin(2.0 * np.pi * (p - np.floor(p)))
-
-    sweep = (x * amp).astype(np.float32)
-    inverse = (x * d * r)[::-1].astype(np.float32)
-    return sweep, inverse
-
-
-def ess_sweep(duration_s: float = 5.0, fmin: float = 20.0,
-              fmax: float = 20000.0, amp: float = 0.5,
-              tail_s: float = 5.0, out_codec: str = 'wav') -> str:
-    import numpy as np
-
-    fmax = min(float(fmax), _AUDIO_RATE * 0.47)
-    sweep, _ = _ess_signals(duration_s, fmin, fmax, amp)
-    tail = np.zeros(int(round(tail_s * _AUDIO_RATE)), dtype=np.float32)
-    mono = np.concatenate([sweep, tail])
-    return _write_wav(np.stack([mono, mono]), out_codec)
-
-
-def deconvolve_ir(recorded_url: str, duration_s: float = 5.0,
-                  fmin: float = 20.0, fmax: float = 20000.0,
-                  amp: float = 0.5, ir_len_s: float = 2.0,
-                  out_codec: str = 'wav') -> str:
-    import numpy as np
-    from scipy.signal import fftconvolve
-
-    rec = _decode_audio_to_array(localize(recorded_url))
-    if rec.shape[1] == 0:
-        raise RuntimeError("deconvolve: recording has no audio")
-    fmax = min(float(fmax), _AUDIO_RATE * 0.47)
-    _, inverse = _ess_signals(duration_s, fmin, fmax, amp)
-
-    ir = np.stack([
-        fftconvolve(rec[c], inverse, mode='full').astype(np.float32)
-        for c in range(2)
-    ])
-    onset = int(np.argmax(np.abs(ir).max(axis=0)))
-    end = min(ir.shape[1], onset + int(round(ir_len_s * _AUDIO_RATE)))
-    ir = ir[:, onset:end]
-    peak = float(np.abs(ir).max())
-    if peak > 0:
-        ir *= 0.99 / peak
-    return _write_wav(ir, out_codec)
+from .audio_render import *  # noqa: F401,F403
+from .audio_ir import *  # noqa: F401,F403
 
 
 __all__ = [
